S-plan batch 9: CI hardening — -Werror clean with native GCC
Fixed 100+ warnings found by x86_64-linux-gnu-gcc (native) with -Werror:
  - Missing forward declarations: struct cpu_context, timer.h includes,
    vfs_open/vfs_close, tpm_is_present
  - Unused static functions: 35+ functions marked __attribute__((unused))
  - Type mismatches: vfs_getxattr 4th arg (int not size_t*),
    kfence_pool_start pointer→uint64_t cast
  - NASM dword overflow: kpti_trampoline.asm -Wno-number-overflow
  - Signedness: mptcp key gen (shift ≥ width), sctp data_len uint16_t→size_t
  - KUnit sentinel: {NULL,NULL} → {0} to include params field
  - Redefines: DT_LNK collision (syscall.h vs readdir.c)
  - Implicit decls: timer_get_ticks, vfs_open, timer_schedule
  - IMA/EVM/IPE/IPE: xattr API misuse (ptr vs value)
  - 6lowpan dispatch mask fix: 0xC0→0xE0 for IPHC detection

Build: zero errors, zero warnings with CC=x86_64-linux-gnu-gcc CFLAGS_EXTRA=-Werror

+++ src/kernel/kfence.c
@@ -151,7 +151,7 @@ void kfence_init(void)
 
     /* Allocate the KFENCE pool from physical memory */
     kfence_pool_size = KFENCE_POOL_SIZE;
-    kfence_pool_start = pmm_alloc_frames(kfence_pool_size / PAGE_SIZE);
+    kfence_pool_start = (uint64_t)pmm_alloc_frames(kfence_pool_size / PAGE_SIZE);
     if (!kfence_pool_start) {
         kprintf("[KFENCE] Failed to allocate pool (%llu bytes)\n",
                 (unsigned long long)kfence_pool_size);
